Movement (#3)
* Made the ships movement have acceleration giving it a
more natural feeling, want to do the same with the enemies as well as updating there movement function to not be so dumb.

* Enemy bullets know hit you and are a different color also refactored
enemy update function to be inside the enemy class

* Started work on a more complex enemy movement algorithm
got the basics working, enemy replusle from one another
and try and move towards the player but need to add more
factors to there movement.

* Further updated the enemy movement function, added

enemies now follow a randomly generated target
that switchs periodically. looks weird by itself
but in combination with a more diverse enemy set
could be a viable algorithm. Also started on an
enemy dodge mechanice as well as a player attraction
mechanice.