Health bar is working and is based off of ships HP
Uses 2 rectangle shapes to represent the players health. I think it would
be nice if a similar concept was used to show the weapons reloading. Also
health is now a float which does make things a lot nicer. There is still a
lot of tweaking needed to be done in terms of game balance.