Fixed the death laser bug, added finer controls
Takes advantage of the range of motion of the joystick and maps the
acceleration to how far the joystick is being pressed. Controls feel great
besides the fact that you cannot shoot the laser and the photon at the
same time but well call that a feature, not  a flaw.